changed imports, moved encoder to subfolder
using absolute imports;
moved encoder to subfolder 'tools';
removed TypeEncoder from __all__
